renamed vendored typing into _typing to prevent issues with Pythohn 2 interpreters
