+++ commands/autoformat.py
@@ -9,7 +9,7 @@ import sublime
 import sublime_plugin
 
 from ..anaconda_lib.worker import Worker
-from ..anaconda_lib.typing import Dict, Any
+from ..anaconda_lib._typing import Dict, Any
 from ..anaconda_lib.progress_bar import ProgressBar
 from ..anaconda_lib.helpers import get_settings, is_python, get_window_view
 from ..anaconda_lib.jsonclient import Callback
