+++ commands/get_lines.py
@@ -5,7 +5,7 @@
 import sublime
 import sublime_plugin
 
-from ..anaconda_lib.typing import Dict, Any
+from ..anaconda_lib._typing import Dict, Any
 from ..anaconda_lib.helpers import get_settings
 from ..anaconda_lib.helpers import valid_languages
 from ..anaconda_lib.linting.sublime import ANACONDA
@@ -35,8 +35,8 @@ class AnacondaGetLines(sublime_plugin.WindowCommand):
         """
 
         view = self.window.active_view()
-        if (view.file_name() in ANACONDA['DISABLED']
-                or not get_settings(view, 'anaconda_linting')):
+        if (view.file_name() in ANACONDA['DISABLED'] or
+                not get_settings(view, 'anaconda_linting')):
             return False
 
         location = view.sel()[0].begin()
